Test writing empty RIFF
